requirements.txt
bwa
samtools
primer3
biopython
